+++ base/arrayshow.jl
@@ -166,7 +166,7 @@ function print_matrix(io::IO, X::AbstractVecOrMat,
     screenwidth -= length(pre) + length(post)
     presp = repeat(" ", length(pre))  # indent each row to match pre string
     postsp = ""
-    @assert textwidth(hdots) == textwidth(ddots)
+    @assert Unicode.textwidth(hdots) == Unicode.textwidth(ddots)
     sepsize = length(sep)
     rowsA, colsA = indices(X,1), indices(X,2)
     m, n = length(rowsA), length(colsA)