+++ base/strings/util.jl
@@ -220,9 +220,9 @@ strip(s::AbstractString, chars::Chars) = lstrip(rstrip(s, chars), chars)
 ## string padding functions ##
 
 function lpad(s::AbstractString, n::Integer, p::AbstractString=" ")
-    m = n - textwidth(s)
+    m = n - Unicode.textwidth(s)
     (m <= 0) && (return s)
-    l = textwidth(p)
+    l = Unicode.textwidth(p)
     if l==1
         return string(p^m, s)
     end
@@ -233,9 +233,9 @@ function lpad(s::AbstractString, n::Integer, p::AbstractString=" ")
 end
 
 function rpad(s::AbstractString, n::Integer, p::AbstractString=" ")
-    m = n - textwidth(s)
+    m = n - Unicode.textwidth(s)
     (m <= 0) && (return s)
-    l = textwidth(p)
+    l = Unicode.textwidth(p)
     if l==1
         return string(s, p^m)
     end