+++ test/arrayops.jl
@@ -467,8 +467,8 @@ end
 @testset "find with general iterables" begin
     s = "julia"
     @test find(c -> c == 'l', s) == [3]
-    g = graphemes("日本語")
-    @test find(isascii, g) == Int[]
+    g = Base.Unicode.graphemes("日本語")
+    @test find(Base.Unicode.isascii, g) == Int[]
     @test find(!iszero, (i % 2 for i in 1:10)) == collect(1:2:9)
 end
 @testset "findn" begin