+++ test/iobuffer.jl
@@ -258,6 +258,7 @@ let io = IOBuffer()
 end
 
 # skipchars
+using Base.Unicode: isspace
 let
     io = IOBuffer("")
     @test eof(skipchars(io, isspace))
@@ -278,7 +279,7 @@ let
 
     for char in ['@','߷','࿊','𐋺']
         io = IOBuffer("alphabeticalstuff$char")
-        @test !eof(skipchars(io, isalpha))
+        @test !eof(skipchars(io, Base.Unicode.isalpha))
         @test read(io, Char) == char
     end
 end